Conflicts:
	proccontrol/h/PCErrors.h
	proccontrol/h/Process.h
	proccontrol/h/ProcessSet.h
	proccontrol/ppc64_bgq_ion/Makefile
	proccontrol/src/bluegeneq.C
	proccontrol/src/int_process.h
	proccontrol/src/linux.C
	proccontrol/src/process.C
	proccontrol/src/procset.C
	proccontrol/src/response.h
	testsuite/amd64-unknown-freebsd7.2/make.solo_mutatee.gen
	testsuite/amd64-unknown-freebsd7.2/tuples
	testsuite/i386-unknown-freebsd7.2/make.solo_mutatee.gen
	testsuite/i386-unknown-freebsd7.2/tuples
	testsuite/i386-unknown-linux2.4/make.mutators.gen
	testsuite/i386-unknown-linux2.4/make.solo_mutatee.gen
	testsuite/i386-unknown-linux2.4/tuples
	testsuite/i386-unknown-nt4.0/tuples
	testsuite/ppc32_bgp/make.solo_mutatee.gen
	testsuite/ppc32_bgp/tuples
	testsuite/ppc32_bgp_ion/make.solo_mutatee.gen
	testsuite/ppc32_bgp_ion/tuples
	testsuite/ppc32_linux/make.solo_mutatee.gen
	testsuite/ppc32_linux/tuples
	testsuite/ppc64_bgq_ion/Makefile
	testsuite/ppc64_bgq_ion/make.mutators.gen
	testsuite/ppc64_bgq_ion/make.solo_mutatee.gen
	testsuite/ppc64_bgq_ion/test_info_new.gen.C
	testsuite/ppc64_bgq_ion/tuples
	testsuite/ppc64_linux/make.solo_mutatee.gen
	testsuite/ppc64_linux/tuples
	testsuite/rs6000-ibm-aix5.1/tuples
	testsuite/src/dynlmon.C
	testsuite/src/proccontrol/pc_groups_mutatee.c
	testsuite/src/proccontrol/proccontrol_comp.C
	testsuite/src/testdriver_be.C
	testsuite/x86_64-unknown-linux2.4/make.mutators.gen
	testsuite/x86_64-unknown-linux2.4/make.solo_mutatee.gen
	testsuite/x86_64-unknown-linux2.4/tuples
	testsuite/x86_64_cnl/make.mutators.gen
	testsuite/x86_64_cnl/make.solo_mutatee.gen
	testsuite/x86_64_cnl/tuples

+++ dyninstAPI_RT/src/RTmemEmulator.c
@@ -187,7 +187,7 @@ unsigned long RTtranslateMemoryShift(unsigned long input, unsigned long origAddr
 int RTuntranslatedEntryCounter;
 
 void RThandleShadow(void *direction, void *pointAddr, void *callbackID, void *flags, void *calculation) {
-    if ((int)direction == 1) {
+   if ((int)((long) direction) == 1) {
         if (RTuntranslatedEntryCounter == 0) {
             // Entering a system call...
             DYNINST_stopThread(pointAddr, callbackID, flags, (void *)1);
